requirements.txt
flask
flask-sqlalchemy
flask-migrate
python-dotenv
pytz
requests
gunicorn
psycopg2-binary
